requirements.txt
langchain==0.1.12
langchain-community==0.0.27
langchain-core==0.1.27
langchain-neo4j==0.0.1
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
pandas==2.2.1
neo4j==5.17.0
python-multipart==0.0.9 